Hide FxA migration banner once dev uses FxA (bug 1080039)
