+++ templates/base.html
@@ -118,7 +118,7 @@
     {% endif %}
     {% if waffle.switch('fx-accounts-migration') and
           (not logged or user.can_migrate_to_fxa()) %}
-      <mkt-banner class="mkt-cloak" success dismiss="off">
+      <mkt-banner id="fxa-migration-banner" class="mkt-cloak" success dismiss="off">
         <span>
           {{ _('Firefox Accounts has arrived') }}
         </span>
